+++ src/lib/intercom/in-process-request-handler.ts
@@ -77,7 +77,7 @@ export async function processInProcessRequest(req: WalletRequest, label: string)
     case WalletMessageType.PrepareRecoveryRequest:
       return {
         type: WalletMessageType.PrepareRecoveryResponse,
-        ready: await Actions.prepareRecoveryTransaction(req.transactionId)
+        ...(await Actions.prepareRecoveryTransaction(req.transactionId))
       };
     case WalletMessageType.ReleaseRecoveryRequest:
       await Actions.releaseRecoveryAuthorization(req.transactionId);